Move some classes out of MessageController
* secretprovider
* template context creation
