+++ src/Runner.Server/Services/Secret.cs
@@ -1,0 +1,8 @@
+namespace Runner.Server.Services
+{
+    public class Secret
+    {
+        public string Name {get;set;}
+        public string Value {get;set;}
+    }
+}
