+++ src/Runner.Server/Services/WorkflowState.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using System.Threading;
+
+namespace Runner.Server.Services
+{
+    public class WorkflowState
+    {
+        public CancellationTokenSource Cancel;
+        public CancellationTokenSource ForceCancel;
+        public Dictionary<string, TaskMetaData> TasksByNameAndVersion;
+    }
+}
