+++ src/Sdk/Pipelines/ExecutionContext.cs
@@ -2,7 +2,7 @@ using System.Threading;
 
 namespace Sdk.Pipelines
 {
-    class ExecutionContext
+    public class ExecutionContext
     {
         public CancellationToken Cancelled { get; set; }
         public JobItemFacade JobContext { get; set; }
